// src/model/User.tsx
export interface User {
    id: number;
    prefix: string;
    firstname: string;
    lastname: string;
    birthday: Date;
    region: string;
    citizen: string;
    nationalId1: string;
    nationalId2: string;
    nationalId3: string;
    nationalId4: string;
    nationalId5: string;
    gender: string;
    prefixNum: string;
    telephone: string;
    passport: string;
    salary: string;
}

export default User;